+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fixed startup order of `Tornex.Scheduler.QueryRegistry` in `Tornex.Scheduler.Supervisor`
+- Fixed `Tornex.Scheduler.Timer` failing to dump queries in other nodes
 
 ## [0.6.0] - 2026-06-05
 ### Added